dsv41: cut the host-side stalls out of the forward
A profiler trace of the log_prob phase showed 21.4 s of a 92 s window with the
GPU idle, and three host-side stalls accounted for 19.8 of it:

- the engram row gather ran on one thread, because ray gives an actor
  OMP_NUM_THREADS=1, and allocated a fresh pinned tensor per call: 13.5 s. It now
  gathers multi-threaded into pinned buffers that are reused across forwards, and
  the table can be mapped whole per rank so the lookup needs no all-reduce.
- the cp_size==1 reference check in cp_utils rebuilt the window index tensor on
  the host and compared it with two device-to-host syncs per layer: 3.8 s.
- deterministic mode's fill of every torch.empty: 2.5 s idle plus 3.4 s of fill
  kernels. MILES_FILL_UNINIT=1 restores it.

The index tensors also build and concatenate in int32, which is what the sparse
attention takes, and the indexer masks its score tensor in place.

GPU busy over the window goes 70.4% -> 85.9%, idle 21.4 s -> 2.8 s, and the
log_prob phase 599 s -> 509 s on the same batch.

+++ miles_plugins/models/deepseek_v41/deepseek_v41.py
@@ -542,7 +542,7 @@ class DeepSeekV41Attention(MegatronModule):
                     rt.candidates = candidates
             topk_rel = rt.topk
             assert topk_rel is not None, f"layer {self.layer_id}: no index_source layer ran before it"
-            compress_idxs = torch.where(topk_rel < 0, -1, topk_rel + offset)
+            compress_idxs = torch.where(topk_rel < 0, -1, topk_rel + offset).int()
             topk_idxs = torch.cat([topk_idxs, compress_idxs], dim=-1)
             kv = torch.cat([kv, latent], dim=1)
